Appling changes at leave management system

+++ HR.LeaveManagement.Application/Features/LeaveAllocation/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveAllocation.Request.Commands;
-using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Application.Contracts.Persistence;
 using MediatR;
 
 
@@ -18,6 +19,10 @@ namespace HR.LeaveManagement.Application.Features.LeaveAllocation.Handlers.Comma
         public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+
+            if (leaveAllocation == null)
+                throw new NotFoundException(nameof(leaveAllocation), request.Id);
+
 
             await _leaveAllocationRepository.Delete(leaveAllocation);
